+++ modules/@apostrophecms/doc-type/index.js
@@ -91,6 +91,7 @@ module.exports = {
     self.composeSchema();
     self.apos.doc.setManager(self.name, self);
     self.enableBrowserData();
+    self.addContextMenu();
   },
   handlers(self) {
     return {
@@ -281,6 +282,16 @@ module.exports = {
 
   methods(self) {
     return {
+      addContextMenu() {
+        self.apos.doc.addContextOperation(self.__meta.name, {
+          action: 'shareDraft',
+          context: 'update',
+          label: 'apostrophe:shareDraft',
+          modal: 'AposModalShareDraft',
+          manuallyPublished: true,
+          hasUrl: true
+        });
+      },
       getRelatedDocsIds(req, doc) {
         const relatedDocsIds = [];
         const handlers = {
@@ -1199,6 +1210,54 @@ module.exports = {
             delete $set[name];
           }
         }
+      },
+
+      async share(req, doc) {
+        if (doc._edit !== true) {
+          throw self.apos.error('notfound');
+        }
+
+        if (!doc._url) {
+          return doc;
+        }
+
+        const { aposShareKey: _aposShareKey, ...draft } = doc;
+        const aposShareKey = doc.aposShareKey || self.apos.util.generateId();
+
+        await self.apos.doc.db.updateOne({
+          _id: doc._id
+        }, {
+          $set: {
+            aposShareKey
+          }
+        });
+
+        return {
+          ...draft,
+          aposShareKey
+        };
+      },
+
+      async unshare(req, doc) {
+        if (doc._edit !== true) {
+          throw self.apos.error('notfound');
+        }
+
+        if (!doc._url) {
+          return doc;
+        }
+
+        const { aposShareKey: _aposShareKey, ...draft } = doc;
+
+        await self.apos.doc.db.updateOne({
+          _id: doc._id
+        }, {
+          $unset: {
+            aposShareKey: 1
+          }
+        });
+
+        return draft;
       }
     };
   },
@@ -1232,6 +1291,28 @@ module.exports = {
   queries(self, query) {
     return {
       builders: {
+        transformDraftForSharing: {
+          after(results) {
+            if (!self.isShareDraftRequest(query.req)) {
+              return;
+            }
+
+            const { aposShareId, aposShareKey } = query.req.query;
+
+            // Change drafts values to make it pass for a published document
+            results.forEach(transformDraftToPublished);
+
+            function transformDraftToPublished (result) {
+              if (result._id === aposShareId && result.aposShareKey === aposShareKey) {
+                const changeToPublished = string => string.replace(':draft', ':published');
+
+                result._id = changeToPublished(result._id);
+                result.aposLocale = changeToPublished(result.aposLocale);
+                result.aposMode = 'published';
+              }
+            }
+          }
+        },
         // `.criteria({...})` Sets the MongoDB criteria, discarding
         // criteria previously added using this
         // method or the `and` method. For this reason,
@@ -2061,16 +2142,26 @@ module.exports = {
               queryLocale = `${query.req.locale}:${query.req.mode}`;
             }
             if (queryLocale) {
-              query.and({
-                $or: [
-                  {
-                    aposLocale: queryLocale
-                  },
-                  {
-                    aposLocale: null
-                  }
-                ]
-              });
+              const $or = [
+                {
+                  aposLocale: queryLocale
+                },
+                {
+                  aposLocale: null
+                }
+              ];
+
+              if (self.isShareDraftRequest(query.req)) {
+                const { aposShareId, aposShareKey } = query.req.query;
+
+                $or.push({
+                  _id: aposShareId,
+                  aposShareKey,
+                  aposLocale: queryLocale.replace(':published', ':draft')
+                });
+              }
+
+              query.and({ $or });
             }
           }
         }