+++ test/docs.js
@@ -166,16 +166,16 @@ describe('Docs', function() {
       label: 'Menu Label',
       modal: 'SomeModalComponent'
     };
-    assert.strictEqual(apos.doc.contextOperations.length, 0);
+    const initialLength = apos.doc.contextOperations.length;
 
     apos.doc.addContextOperation('test-people', operation);
-    assert.strictEqual(apos.doc.contextOperations.length, 1);
-    assert.deepStrictEqual(apos.doc.contextOperations[0], {
+
+    assert.strictEqual(apos.doc.contextOperations.length, initialLength + 1);
+    assert.deepStrictEqual(apos.doc.contextOperations.find(op => op.action === 'test'), {
       ...operation,
       moduleName: 'test-people'
     });
   });
-
   it('should support custom context menu (with optional)', async function() {
     apos.doc.contextOperations = [];
     const operation = {